requirements.txt
flask==2.3.3
langchain==0.1.0
langchain-openai==0.0.5
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
openai==1.3.7
tiktoken==0.5.2
requests==2.31.0